created delete todo popup

+++ src/app/shared/components/todo-row/todo-row.component.ts
@@ -3,15 +3,19 @@ import {
   booleanAttribute,
   ChangeDetectionStrategy,
   Component,
+  DestroyRef,
   inject,
   input,
   InputSignal,
   InputSignalWithTransform
 } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { MatDialog } from '@angular/material/dialog';
 import { MatIcon } from '@angular/material/icon';
 import { Debounced } from '../../decorators';
 import { Todo } from '../../interfaces';
 import { TodosService } from '../../services';
+import { DeleteTodoPopupComponent } from '../delete-todo-popup';
 import { TimerComponent } from '../timer';
 
 @Component({
@@ -24,6 +28,8 @@ import { TimerComponent } from '../timer';
 })
 export class TodoRowComponent {
   readonly #todosService = inject(TodosService);
+  readonly #dialog = inject(MatDialog);
+  readonly #destroyRef = inject(DestroyRef);
 
   todo: InputSignal<Todo> = input.required();
   isToday: InputSignalWithTransform<boolean, unknown> = input(false, { transform: booleanAttribute });
@@ -34,6 +40,14 @@ export class TodoRowComponent {
   }
 
   protected handleDeleteTodo(): void {
-    this.#todosService.delete(this.todo().id)
+    this.#dialog.open(DeleteTodoPopupComponent, {
+      data: this.todo().title
+    })
+      .afterClosed()
+      .pipe(takeUntilDestroyed(this.#destroyRef))
+      .subscribe((isConfirmed?: boolean) => {
+        if (!isConfirmed) return;
+        this.#todosService.delete(this.todo().id);
+      });
   }
 }
